requirements.txt
customtkinter==5.2.2
tkcalendar==1.6.1
Pillow==10.2.0
pyinstaller==6.11.1 